# docs/swagger.yaml
components:
  schemas:
...
    analysis.CapabilityManifest:
      properties:
        commands:
          items:
            $ref: '#/components/schemas/analysis.Capability'
          type: array
          uniqueItems: false
        documentation_snapshot:
          type: string
        functions:
          items:
            $ref: '#/components/schemas/analysis.Capability'
          type: array
          uniqueItems: false
        language:
          type: string
        profile:
          type: string
        rewrite:
          $ref: '#/components/schemas/analysis.RewriteCapabilityManifest'
          description: Optional rewrite support for this selected dialect; inspect
            each form rather than assuming universal support.
        schema_version:
          type: integer
        version:
          type: string
      type: object
    analysis.Coverage:
      properties:
        reasons:
          items:
            type: string
          type: array
          uniqueItems: false
        semantic_complete:
          type: boolean
        syntax_complete:
          type: boolean
      type: object
...
    analysis.Result:
      properties:
        coverage:
          $ref: '#/components/schemas/analysis.Coverage'
        dependencies:
          $ref: '#/components/schemas/analysis.Dependencies'
        diagnostics:
          items:
            $ref: '#/components/schemas/analysis.Diagnostic'
          type: array
          uniqueItems: false
        document:
          $ref: '#/components/schemas/analysis.QueryDocument'
        lineage:
          items:
            $ref: '#/components/schemas/analysis.Lineage'
          type: array
          uniqueItems: false
        references:
          items:
            $ref: '#/components/schemas/analysis.Reference'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        scopes:
          items:
            $ref: '#/components/schemas/analysis.Scope'
          type: array
          uniqueItems: false
        stages:
          items:
            $ref: '#/components/schemas/analysis.Stage'
          type: array
          uniqueItems: false
        status:
          $ref: '#/components/schemas/analysis.Status'
      type: object
...
    api.AnalysisRequest:
      additionalProperties: false
      properties:
        language:
          default: spl
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - spl
          - spl2
          type: string
        profile:
          default: splunkd
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - splunkd
          type: string
        source_id:
          type: string
        text:
          type: string
        version:
          default: current
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - current
          type: string
      required:
      - text
      type: object
    api.DiscoverQueryRequest:
      description: Request to discover information about an SPL query
      properties:
        language:
          description: spl (default); spl2 is rejected for this legacy operation
          type: string
        profile:
          description: splunkd (default)
          type: string
        query:
          description: SPL query to analyze
          example: search sourcetype=access_combined | stats count by src_ip
          type: string
        version:
          description: current (default)
          type: string
      required:
      - query
      type: object
    api.DiscoverQueryResponse:
      description: Response from discovering SPL query information
      properties:
        query:
          description: Original query
          example: search sourcetype=access_combined | stats count by src_ip
          type: string
          x-order: '1'
        query_info:
          $ref: '#/components/schemas/mapper.QueryInfo'
        success:
          description: Whether the discovery was successful
          example: true
          type: boolean
          x-order: '3'
      type: object
    api.ErrorResponse:
      description: Standard error response
      properties:
        code:
          description: HTTP status code
          example: 400
          type: integer
          x-order: '3'
        error:
          description: Always true for error responses
          example: true
          type: boolean
          x-order: '1'
        message:
          description: Human-readable error message
          example: Invalid request
          type: string
          x-order: '2'
      type: object
...
    api.ValidationErrorResponse:
      description: Response with detailed validation errors
      properties:
        code:
          description: HTTP status code
          example: 400
          type: integer
          x-order: '3'
        error:
          description: Always true for error responses
          example: true
          type: boolean
          x-order: '1'
        message:
          description: General error message
          example: Validation failed
          type: string
          x-order: '2'
        validation_errors:
          description: List of specific validation errors
          items:
            $ref: '#/components/schemas/api.ValidationError'
          type: array
          uniqueItems: false
          x-order: '4'
      type: object
...
info:
  contact:
    email: support@example.com
    name: SPL Toolkit Support
  description: Offline field mapping, flat discovery, and validation for the supported
    Milestone 1 SPL grammar.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SPL Toolkit API
  version: 1.0.0
openapi: 3.1.0
paths:
  /capabilities:
    get:
      description: Select the canonical capability manifest including language, profile,
        compatibility version, commands, functions, limitations, and optional pinned
        documentation_snapshot. Only language/profile/version query parameters are
        accepted; duplicate keys are rejected even when equal.
      parameters:
      - description: 'Query language: spl (default) or spl2; empty uses default'
        in: query
        name: language
        schema:
          enum:
          - ''
          - spl
          - spl2
          type: string
      - description: 'Execution profile: splunkd (default); empty uses default'
        in: query
        name: profile
        schema:
          enum:
          - ''
          - splunkd
          type: string
      - description: 'Compatibility version: current (default); empty uses default'
        in: query
        name: version
        schema:
          enum:
          - ''
          - current
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/analysis.CapabilityManifest'
          description: Analysis capability manifest
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unknown, duplicate, conflicting, malformed, or unsupported
            selectors
      summary: Get analysis capabilities
      tags:
      - query
...
  /query/analyze:
    post:
      description: Strict query document requiring text, with optional language (spl
        or spl2), profile (splunkd), version (current), and source_id strings. Empty
        selectors use defaults spl/splunkd/current. Preserve text and identity exactly.
        Reject null, duplicate or unknown members, malformed Unicode, and trailing
        JSON. Body limit is 1 MiB. Return canonical coverage, located references,
        SQL execution phases, lineage, dependencies and diagnostics; all content statuses
        use 200.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.AnalysisRequest'
        description: Query document
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/analysis.Result'
          description: Canonical analysis report for valid, invalid, and incomplete
            queries
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Malformed JSON or unsupported document options
      summary: Analyze an SPL or standalone SPL2 query
      tags:
      - query
  /query/discover:
    post:
      description: Analyze an SPL query to discover metadata like datamodels, lookups,
        macros, sources, and input fields
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.DiscoverQueryRequest'
        description: Query discovery request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.DiscoverQueryResponse'
          description: Successfully discovered query information
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ValidationErrorResponse'
          description: Invalid request or validation errors
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Discover information about an SPL query
      tags:
      - query
  /query/map:
    post:
      description: Apply field mappings to transform field names in an SPL query.
        Supports both stateless operation (provide mappings/config in request) and
        stateful operation (use pre-loaded mappings). For production use, provide
        mappings or config in each request to avoid global state.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.MapQueryRequest'
        description: Query mapping request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.MapQueryResponse'
          description: Successfully mapped the query
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ValidationErrorResponse'
          description: Invalid request or validation errors
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Map fields in an SPL query
      tags:
      - query
  /query/rewrite:
    post:
      description: Strict versioned request with one query document, explicit inline
        rules, optional inline validation target, and preview mode by default. No
        filesystem paths, URL retrieval, or mapper configuration. Query statuses valid,
        invalid, and incomplete all return HTTP 200. Body limit is 8 MiB.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.RewriteRequest'
        description: Versioned single-query rewrite request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rewrite.Result'
          description: Canonical rewrite report
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Preview or apply explicit safe rewrite rules
      tags:
      - query
  /query/rewrite/batch:
    post:
      description: Strict versioned request with nonempty query documents, explicit
        inline rules, optional inline validation target, and preview mode by default.
        No filesystem paths, URL retrieval, or mapper configuration. Query statuses
        valid, invalid, and incomplete all return HTTP 200. Body limit is 8 MiB.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.RewriteBatchRequest'
        description: Versioned ordered-batch rewrite request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rewrite.BatchResult'
          description: Canonical ordered rewrite reports
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Preview or apply explicit safe rewrite rules to an ordered batch
      tags:
      - query
  /query/validate:
    post:
      description: Check whether a query is accepted by the bundled Milestone 1 SPL
        grammar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.ValidateQueryRequest'
        description: Query validation request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ValidateQueryResponse'
          description: Query is valid
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ValidationErrorResponse'
          description: Invalid request structure
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ValidateQueryResponse'
          description: Query has invalid syntax
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Validate an SPL query
      tags:
      - query
  /query/validate-fields:
    post:
      description: Strict object with required document and catalog keys. Document
        requires string text and permits only language/profile/version/source_id string
        options. Catalog accepts a string array or an object requiring fields and
        permitting optional_fields/identity/version. Nulls, unknown or duplicate keys,
        malformed Unicode, duplicate names and ordinary/optional overlap are rejected.
        Catalog metadata never resolves files or URLs. Body limit is 1 MiB. Content
        statuses valid/invalid/incomplete all return 200 with the full canonical report.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/validation.Request'
        description: Query document and catalog (catalog may also be a string array)
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation.Report'
          description: Canonical report, including invalid or incomplete content
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Validate query fields against a local catalog
      tags:
      - query
  /query/validate-fields/batch:
    post:
      description: Strict object requiring only documents and catalog. Documents is
        a nonempty array of strict query documents, each with its own options and
        identity. Catalog accepts a string array or an object requiring fields and
        permitting optional_fields/identity/version; names are case-sensitive. Reject
        nulls, unknown or duplicate keys and malformed Unicode. No catalog file or
        URL resolution. Body limit is 1 MiB. Batch status precedence is invalid, incomplete,
        valid; all content statuses return 200.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/validation.BatchRequest'
        description: Nonempty query document array and catalog (catalog may also be
          a string array)
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation.BatchReport'
          description: Canonical ordered batch report
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Validate a batch of query documents against a local catalog
      tags:
      - query
  /query/validate-schema:
    post:
      description: 'Strict wrapper requiring document (or nonempty documents for batch)
        and target. Query documents require text. Reject unknown/duplicate wrapper
        members, nulls and malformed Unicode. Target is a tagged json_schema or ocsf
        union: JSON Schema accepts object-or-boolean schema, optional base_uri and
        URI-keyed inline resources; OCSF requires an official compiled catalog object
        and exact selection version with one of class/class_uid/category/category_uid,
        optional unique profiles/extensions arrays (not null). Schema/catalog payloads
        allow their own vocabulary. No filesystem lookup, downloading or URL dereference.
        Body limit is 8 MiB. Valid, invalid and incomplete content all return 200
        with the full canonical report; input errors return 400.'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.SchemaValidationRequest'
        description: Query document and explicitly supplied schema target
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation.SchemaReport'
          description: Canonical report, including invalid or incomplete content
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Validate query fields against an offline JSON Schema or OCSF target
      tags:
      - query
  /query/validate-schema/batch:
    post:
      description: 'Strict wrapper requiring document (or nonempty documents for batch)
        and target. Query documents require text. Reject unknown/duplicate wrapper
        members, nulls and malformed Unicode. Target is a tagged json_schema or ocsf
        union: JSON Schema accepts object-or-boolean schema, optional base_uri and
        URI-keyed inline resources; OCSF requires an official compiled catalog object
        and exact selection version with one of class/class_uid/category/category_uid,
        optional unique profiles/extensions arrays (not null). Schema/catalog payloads
        allow their own vocabulary. No filesystem lookup, downloading or URL dereference.
        Body limit is 8 MiB. Valid, invalid and incomplete content all return 200
        with the full canonical report; input errors return 400.'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.SchemaValidationBatchRequest'
        description: Nonempty query document array and explicitly supplied schema
          target
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validation.SchemaBatchReport'
          description: Canonical ordered batch report
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Validate an ordered batch against an offline JSON Schema or OCSF target
      tags:
      - query
servers:
- description: Development server
  url: http://localhost:8080/api/v1
...